comment delete function and some style stuff

+++ blogs/views.py
@@ -9,6 +9,7 @@ from django.db.models import Q
 from django.contrib.contenttypes.models import ContentType
 from comments.models import Comment
 from comments.forms import CommentForm
+from .utils import count_time
 # Create your views here.
 
 
@@ -97,9 +98,9 @@ def posts_list(request):
 
 
 def posts_update(request, id):
-    if not request.user.is_authenticated:
-        raise Http404
     instance = get_object_or_404(Post, id=id)
+    if instance.user != request.user or not request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -118,6 +119,11 @@ def posts_delete(request,id):
     if not request.user.is_staff or not request.user.is_superuser:
         raise Http404
     instance = get_object_or_404(Post, id=id)
-    instance.delete()
-    messages.success(request, "Successfully Deleted")
-    return redirect("posts_list")
+    if request.method == "POST":
+        instance.delete()
+        messages.success(request, "Successfully Deleted")
+        return redirect("posts_list")
+    context = {
+        "object": instance,
+    }
+    return render(request, "blogs/confirm_delete.html", context)
